cdk/socket-server/requirements.txt
python-socketio[asyncio_client]
boto3
smithy-aws-core==0.0.3
aws_sdk_bedrock_runtime==0.0.2
botocore>=1.34.0
websockets
